+++ app/src/main/AndroidManifest.xml
@@ -30,7 +30,7 @@
 
 
         <activity
-            android:name=".MainMenu"
+            android:name=".ui.general.MainMenu"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.VerseVerwebt"
@@ -41,20 +41,20 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ChapterIntro" android:screenOrientation="portrait" android:resizeableActivity="true" />
-        <activity android:name=".Chapter1" android:screenOrientation="portrait" android:resizeableActivity="true" />
-        <activity android:name=".Chapter2" android:screenOrientation="portrait" android:resizeableActivity="true" />
-        <activity android:name=".Chapter3" android:screenOrientation="portrait" android:resizeableActivity="true" />
-        <activity android:name=".Chapter4" android:screenOrientation="portrait" android:resizeableActivity="true" />
-        <activity android:name=".Chapter5" android:screenOrientation="portrait" android:resizeableActivity="true" />
-        <activity android:name=".Chapter6" android:screenOrientation="portrait" android:resizeableActivity="true" />
-        <activity android:name=".Chapter7" android:screenOrientation="portrait" android:resizeableActivity="true" />
-        <activity android:name=".TableOfContents" android:screenOrientation="portrait" android:resizeableActivity="true"/>
-        <activity android:name=".Ranking" android:screenOrientation="portrait" android:resizeableActivity="true" />
-        <activity android:name=".Login" android:screenOrientation="portrait" android:resizeableActivity="true" />
-        <activity android:name=".SignUp" android:screenOrientation="portrait" android:resizeableActivity="true" />
-        <activity android:name=".Profile" android:screenOrientation="portrait" android:resizeableActivity="true" />
-        <activity android:name=".Credits" android:screenOrientation="portrait" android:resizeableActivity="true" />
+        <activity android:name=".ui.chapters.ChapterIntro" android:screenOrientation="portrait" android:resizeableActivity="true" />
+        <activity android:name=".ui.chapters.Chapter1" android:screenOrientation="portrait" android:resizeableActivity="true" />
+        <activity android:name=".ui.chapters.Chapter2" android:screenOrientation="portrait" android:resizeableActivity="true" />
+        <activity android:name=".ui.chapters.Chapter3" android:screenOrientation="portrait" android:resizeableActivity="true" />
+        <activity android:name=".ui.chapters.Chapter4" android:screenOrientation="portrait" android:resizeableActivity="true" />
+        <activity android:name=".ui.chapters.Chapter5" android:screenOrientation="portrait" android:resizeableActivity="true" />
+        <activity android:name=".ui.chapters.Chapter6" android:screenOrientation="portrait" android:resizeableActivity="true" />
+        <activity android:name=".ui.chapters.Chapter7" android:screenOrientation="portrait" android:resizeableActivity="true" />
+        <activity android:name=".ui.general.TableOfContents" android:screenOrientation="portrait" android:resizeableActivity="true"/>
+        <activity android:name=".ui.user.Ranking" android:screenOrientation="portrait" android:resizeableActivity="true" />
+        <activity android:name=".ui.user.Login" android:screenOrientation="portrait" android:resizeableActivity="true" />
+        <activity android:name=".ui.user.SignUp" android:screenOrientation="portrait" android:resizeableActivity="true" />
+        <activity android:name=".ui.user.Profile" android:screenOrientation="portrait" android:resizeableActivity="true" />
+        <activity android:name=".ui.general.Credits" android:screenOrientation="portrait" android:resizeableActivity="true" />
     </application>
 
 </manifest>
