+++ app/src/main/java/com/example/verseverwebt/ui/chapters/Chapter7.kt
@@ -1,4 +1,4 @@
-package com.example.verseverwebt
+package com.example.verseverwebt.ui.chapters
 
 import androidx.activity.compose.setContent
 import android.database.ContentObserver
@@ -21,6 +21,10 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import com.example.verseverwebt.ui.general.AnimatedTypewriterText2
+import com.example.verseverwebt.ui.general.BackToMenuButton
+import com.example.verseverwebt.ui.general.PageNumber
+import com.example.verseverwebt.R
 import com.example.verseverwebt.ui.theme.CustomTypography
 
 class Chapter7 : ComponentActivity() {
