+++ app/src/main/java/com/example/verseverwebt/ui/chapters/ChapterIntro.kt
@@ -1,4 +1,4 @@
-package com.example.verseverwebt
+package com.example.verseverwebt.ui.chapters
 
 import android.content.Context
 import android.hardware.camera2.CameraManager
@@ -15,8 +15,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.example.verseverwebt.ui.general.AnimatedFadeInText
+import com.example.verseverwebt.ui.general.AnimatedTypewriterText
+import com.example.verseverwebt.ui.general.BackToMenuButton
 import com.example.verseverwebt.api.ApiClient
 import com.example.verseverwebt.ui.theme.CustomTypography
 import com.example.verseverwebt.ui.theme.VerseVerwebtTheme
