+++ app/src/main/java/com/example/verseverwebt/ui/general/Credits.kt
@@ -1,11 +1,10 @@
-package com.example.verseverwebt
+package com.example.verseverwebt.ui.general
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
