+++ app/src/main/java/com/example/verseverwebt/ui/user/Ranking.kt
@@ -1,4 +1,4 @@
-package com.example.verseverwebt
+package com.example.verseverwebt.ui.user
 
 import android.os.Bundle
 import android.util.Log
@@ -13,6 +13,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import com.example.verseverwebt.ui.general.BackToMenuButton
 import com.example.verseverwebt.api.ApiClient
 import com.example.verseverwebt.ui.theme.VerseVerwebtTheme
 import com.example.verseverwebt.ui.theme.CustomTypography
